% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mjs.R
\name{theme_mjs}
\alias{theme_mjs}
\title{A ggplot2 theme for the Milwaukee Journal Sentinel}
\usage{
theme_mjs()
}
\description{
Add this to a ggplot2 plot to apply the default Milwaukee Journal Sentinel style
}
\examples{
ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point() + theme_mjs()
}
\keyword{ggplot2,}
\keyword{theme}
